Feat/messagesfront

+++ Alderto.Services/Impl/MessagesManager.cs
@@ -1,10 +1,10 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Alderto.Data;
 using Alderto.Data.Models;
-using Alderto.Services.Exceptions.BadRequest;
-using Alderto.Services.Exceptions.NotFound;
+using Alderto.Services.Exceptions;
 using Discord;
 using Microsoft.EntityFrameworkCore;
 
@@ -26,59 +26,81 @@ namespace Alderto.Services.Impl
             return _context.GuildManagedMessages.Where(m => m.GuildId == guildId).ToListAsync();
         }
 
-        public async Task<IMessage> GetMessageAsync(ulong guildId, ulong messageId)
+        public async Task<GuildManagedMessage?> GetMessageAsync(ulong guildId, ulong messageId)
         {
-            return await GetDiscordBotMessageAsync(await GetManagedMessage(guildId, messageId));
+            return await _context.GuildManagedMessages.FindAsync(guildId, messageId);
         }
 
-        public async Task<IUserMessage> PostMessageAsync(ulong guildId, ulong channelId, string message)
+        public async Task<GuildManagedMessage> PostMessageAsync(ulong guildId, ulong channelId, string message, ulong? moderatorRoleId = null)
         {
-            var discordMessage = await (await GetDiscordChannel(guildId, channelId)).SendMessageAsync(message);
+            var channel = await GetDiscordChannel(guildId, channelId);
+            var discordMessage = await channel.SendMessageAsync(message);
 
-            _context.GuildManagedMessages.Add(new GuildManagedMessage(guildId, channelId, discordMessage.Id));
+            // Always will result in a new entry, as bot is making a new post.
+            var msg = new GuildManagedMessage(guildId, channelId, discordMessage.Id, discordMessage.Content, moderatorRoleId: moderatorRoleId);
+
+            _context.GuildManagedMessages.Add(msg);
             await _context.SaveChangesAsync();
 
-            return discordMessage;
+            return msg;
         }
 
-        public async Task<IUserMessage> ImportMessageAsync(ulong guildId, ulong channelId, ulong messageId)
+        public async Task<GuildManagedMessage> ImportMessageAsync(ulong guildId, ulong channelId, ulong messageId, ulong? moderatorRoleId = null)
         {
+            var dbMsg = await _context.GuildManagedMessages.SingleOrDefaultAsync(m =>
+                m.GuildId == guildId && m.MessageId == messageId);
+            if (dbMsg != null)
+                return dbMsg;
+
+            // Massage was not yet imported.
             var discordMessage = await GetDiscordBotMessageAsync(guildId, channelId, messageId);
 
-            if (!await _context.GuildManagedMessages.AnyAsync(m => m.GuildId == guildId && m.MessageId == messageId))
-            {
-                _context.GuildManagedMessages.Add(new GuildManagedMessage(guildId, channelId, messageId));
-                await _context.SaveChangesAsync();
-            }
+            _context.GuildManagedMessages.Add(dbMsg = new GuildManagedMessage(guildId, channelId, messageId,
+                discordMessage.Content, discordMessage.EditedTimestamp ?? discordMessage.CreatedAt, moderatorRoleId));
+            await _context.SaveChangesAsync();
 
-            return discordMessage;
+            return dbMsg;
         }
 
-        public async Task EditMessageAsync(ulong guildId, ulong messageId, string newMessageContents)
+        public async Task EditMessageAsync(ulong guildId, ulong messageId, string? content = null, ulong? moderatorRoleId = null)
         {
-            var message = await GetDiscordBotMessageAsync(await GetManagedMessage(guildId, messageId));
+            var dbMsg = await GetMessageAsync(guildId, messageId);
+            if (dbMsg == null)
+                throw new MessageNotFoundException();
+
+            var message = await GetDiscordBotMessageAsync(dbMsg);
 
             // User can always edit its own posts.
-            await message.ModifyAsync(msg => msg.Content = newMessageContents);
+            if (content != null)
+            {
+                await message.ModifyAsync(msg => msg.Content = content);
+                dbMsg.Content = content;
+            }
+
+            if (moderatorRoleId != null)
+                // If edited, check if was edited to not be removed.
+                dbMsg.ModeratorRoleId = moderatorRoleId == 0 ? null : moderatorRoleId;
+
+            dbMsg.LastModified = DateTime.UtcNow;
+
+            await _context.SaveChangesAsync();
         }
 
         public async Task RemoveMessageAsync(ulong guildId, ulong messageId)
         {
-            var message = await GetDiscordBotMessageAsync(await GetManagedMessage(guildId, messageId));
+            var dbMsg = await GetMessageAsync(guildId, messageId);
+            if (dbMsg == null)
+                throw new MessageNotFoundException();
 
-            var dbMsg = await _context.GuildManagedMessages.FindAsync(guildId, messageId);
-            _context.GuildManagedMessages.Remove(dbMsg);
-            await _context.SaveChangesAsync();
+            var message = await GetDiscordBotMessageAsync(dbMsg);
 
             // User can always delete its own posts.
+
+            // First attempt to delete from the server.
             await message.DeleteAsync();
-        }
 
-
-        private async Task<GuildManagedMessage?> GetManagedMessage(ulong guildId, ulong messageId)
-        {
-            return await _context.GuildManagedMessages.SingleOrDefaultAsync(message =>
-                message.GuildId == guildId && message.MessageId == messageId);
+            _context.GuildManagedMessages.Remove(dbMsg);
+            await _context.SaveChangesAsync();
         }
 
         private async Task<IMessageChannel> GetDiscordChannel(ulong guildId, ulong channelId)
@@ -97,11 +119,8 @@ namespace Alderto.Services.Impl
             return msgChannel;
         }
 
-        private Task<IUserMessage> GetDiscordBotMessageAsync(GuildManagedMessage? msg)
+        private Task<IUserMessage> GetDiscordBotMessageAsync(GuildManagedMessage msg)
         {
-            if (msg == null)
-                throw new MessageNotFoundException();
-
             return GetDiscordBotMessageAsync(msg.GuildId, msg.ChannelId, msg.MessageId);
         }
